Turn function for getting audio info into module.
